refactor(canton): extract request_id.rs from indexer_canton
